Added MainMenu, Added Animations, Added Battletrigger
Mainmenu: Added a simple menu placeholder with 3 buttons, two of which work already.

Animations: Added a loading screen (ISNT ACTUAL LOADING SCREEN BUT IS TO HIDE OTHER TRANSITIONS)
Added a transition to the BattleScene

Battletrigger: There is now a layer called 'Battletrigger' which tiles can be placed on. When the player interacts with these tiles, the player has a fixed chance to encounter an enemy (see player script). Still have to fetch information about the battle from the scene the player comes from (to get a fair encounter idea like pokemon).